Add support for palette, reorg the code
JIRA: RAIL-4069
