+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/geoChart/GeoPushpinChartDescriptor.ts
@@ -10,7 +10,7 @@ import {
 } from "@gooddata/sdk-model";
 import { ISettings } from "@gooddata/sdk-backend-spi";
 import { BucketNames } from "@gooddata/sdk-ui";
-import { IGeoConfig, IGeoPushpinChartProps } from "@gooddata/sdk-ui-geo";
+import { IGeoPushpinChartProps } from "@gooddata/sdk-ui-geo";
 
 import {
     IVisualizationDescriptor,
@@ -24,24 +24,11 @@ import { BaseChartDescriptor } from "../baseChart/BaseChartDescriptor";
 import {
     bucketConversion,
     chartAdditionalFactories,
-    getConfigFromPropsConverter,
     getInsightToPropsConverter,
     getReactEmbeddingCodeGenerator,
     insightConversion,
 } from "../../../utils/embeddingCodeGenerator";
-
-const supportedGeoConfigProperties = new Set<keyof IGeoConfig>([
-    "center",
-    "colorMapping",
-    "cooperativeGestures",
-    "legend",
-    "limit",
-    "selectedSegmentItems",
-    "separators",
-    "viewport",
-    "points",
-    "showLabels",
-]);
+import { geoConfigFromInsight } from "./geoConfigFromInsight";
 
 export class GeoPushpinChartDescriptor extends BaseChartDescriptor implements IVisualizationDescriptor {
     public getFactory(): PluggableVisualizationFactory {
@@ -80,7 +67,7 @@ export class GeoPushpinChartDescriptor extends BaseChartDescriptor implements IV
             segmentBy: bucketConversion("segmentBy", BucketNames.SEGMENT, bucketAttribute),
             filters: insightConversion("filters", insightFilters),
             sortBy: insightConversion("sortBy", insightSorts),
-            config: insightConversion("config", getConfigFromPropsConverter(supportedGeoConfigProperties)),
+            config: insightConversion("config", geoConfigFromInsight),
         }),
         additionalFactories: chartAdditionalFactories,
     });
