+++ libs/sdk-ui-ext/src/internal/interfaces/VisualizationDescriptor.ts
@@ -1,6 +1,6 @@
 // (C) 2021-2022 GoodData Corporation
 import { IAnalyticalBackend, ISettings, IUserWorkspaceSettings } from "@gooddata/sdk-backend-spi";
-import { IInsight, IInsightDefinition } from "@gooddata/sdk-model";
+import { IColorPalette, IInsight, IInsightDefinition } from "@gooddata/sdk-model";
 import { IFluidLayoutDescriptor } from "./LayoutDescriptor";
 import { IDrillDownContext, IVisConstruct, IVisualization } from "./Visualization";
 
@@ -45,11 +45,18 @@ export interface IEmbeddingCodeContext {
      * If not provided, any backend-dependent properties MUST be ignored in the code generator.
      */
     backend?: IAnalyticalBackend;
+
     /**
      * Settings of the current user.
      * If not provided, any user specific-settings MUST be ignored in the code generator.
      */
     settings?: IUserWorkspaceSettings;
+
+    /**
+     * The color palette to use when rendering the visualization.
+     * If not provided, the default color palette will be used.
+     */
+    colorPalette?: IColorPalette;
 }
 
 /**
