+++ libs/sdk-ui-ext/src/internal/utils/embeddingCodeGenerator/index.ts
@@ -1,10 +1,4 @@
 // (C) 2022 GoodData Corporation
 export { getReactEmbeddingCodeGenerator } from "./getReactEmbeddingCodeGenerator";
-export {
-    bucketConversion,
-    insightConversion,
-    chartConfigFromInsight,
-    getConfigFromPropsConverter,
-    getInsightToPropsConverter,
-} from "./insightToPropsConverter";
+export { bucketConversion, insightConversion, getInsightToPropsConverter } from "./insightToPropsConverter";
 export { chartAdditionalFactories } from "./chartAdditionalFactories";
