+++ libs/sdk-ui-ext/src/internal/utils/tests/removeUseless.test.ts
